+++ common/src/main/java/net/smackplays/smacksutil/CommonClass.java
@@ -17,7 +17,9 @@ public class CommonClass {
 
         Constants.LOG.info("Hello from Common init on {}! we are currently in a {} environment!", Services.PLATFORM.getPlatformName(), Services.PLATFORM.getEnvironmentName());
         Constants.LOG.info("The ID for diamonds is {}", BuiltInRegistries.ITEM.getKey(Items.DIAMOND));
-        Services.CONFIG.init();
+        if (Services.CONFIG != null) {
+            Services.CONFIG.init();
+        }
         // It is common for all supported loaders to provide a similar feature that can not be used directly in the
         // common code. A popular way to get around this is using Java's built-in service loader feature to create
         // your own abstraction layer. You can learn more about this in our provided services class. In this example
