+++ fabric/src/main/java/net/smackplays/smacksutil/screens/BackpackScreen.java
@@ -7,6 +7,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
 import net.smackplays.smacksutil.ModClient;
+import net.smackplays.smacksutil.SmacksUtil;
 import net.smackplays.smacksutil.menus.BackpackMenu;
 
 
@@ -20,8 +21,8 @@ public class BackpackScreen extends AbstractBackpackScreen<BackpackMenu> {
         FriendlyByteBuf packet = PacketByteBufs.create();
         ItemStack stack = this.menu.playerInventory.getSelected();
         packet.writeItem(stack);
-        if (ClientPlayNetworking.canSend(ModClient.SORT_REQUEST_ID)){
-            ClientPlayNetworking.send(ModClient.SORT_REQUEST_ID, packet);
+        if (ClientPlayNetworking.canSend(SmacksUtil.SORT_REQUEST_ID)){
+            ClientPlayNetworking.send(SmacksUtil.SORT_REQUEST_ID, packet);
         }
     }
 }
